fix chat messages not always being sorted
I think it had to do with responses to HTTP request for chat messages
arriving out-of-order, whereas the code depended on them arriving
in-order.
This is a sledgehammer approach, but it works and shouldn't be too
costly looking at the number of chat messages and how infrequent this
function is called.

Fixes #12