controls: don't render when invisible
On a phone you might tap the slider without even being able to see it
(pointer events are still captured while it is invisible).

I originally did not do this because I thought rendering it in the
background might lead to a little more performance when going to show
it.  But in actuality it might be the other way around, Netflix only
renders it when needed, for example.