introDurations: fix stupid logic error
When watching a stream which has a einde LW from time 0, the intro
button appears but doesn't really skip to anything useful.
This button that appears at time [0, introDuration] is meant for streams
without a einde LW.
Make this explicit so the button only shows up when it does something
remotely useful.